requirements.txt
google-adk
google-adk[otel-gcp]
google-genai
google-cloud-storage
toolbox-core
asyncpg